build: replace gif-resize usage with sharp
In addition I was able to convert gifs to webp files now, which
simplified the generate-responsive-imgs script. I also refactored the
script into more discreet functions to make it a bit easier to read.
Now it should be more clear about the special cases that happen in the
script.

Also .eleventy.js was updated to convert gifs, jpgs, and pngs
to webm in the built site.